requirements.txt
audioread==3.0.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
fonttools==4.59.0
idna==3.10
joblib==1.5.1
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
matplotlib==3.10.5
msgpack==1.1.1
numba==0.61.2
numpy==2.2.6
packaging==25.0
pandas==2.3.1
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==6.2.0
requests==2.32.4
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
spotipy==2.25.1
threadpoolctl==3.6.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
voyager==2.1.0
